Make parser lenient for list and map record fields.
This might be a bit controvertial, as usually defaulting things to
be 'helpful' is frought. In this case though, I think it makes
sense for backwards compatibility to allow list fields have no
value if the field is empty.

Note that things like `Option[List[A]]` still _do_ work as expected,
as the special case is not transitive.